Dodaj float i  doubl nie wiem czy nei tzreba zmienic int an float. program.cs ładny. reszta w ostatnim commit

+++ StudentJournal/StudentJournal/IStudent.cs
@@ -7,11 +7,11 @@
         string Age { get; }
 
 
-        void AddGradeMath(float grade);
-        void AddGradePolish(float grade);
-        void AddGradeEnglish(float grade);
-        void AddGradeIT(float grade);
-        void AddGradePhysics(float grade);
+        void AddGradeMath(int grade);
+        void AddGradePolish(int grade);
+        void AddGradeEnglish(int grade);
+        void AddGradeIT(int grade);
+        void AddGradePhysics(int grade);
         void AddGradeMath(string grade);
         void AddGradePolish(string grade);
         void AddGradeEnglish(string grade);
@@ -22,6 +22,7 @@
         void AddGradeEnglish(char grade);
         void AddGradeIT(char grade);
         void AddGradePhysics(char grade);
+        
 
         Statistics GetStatisticsMath();
         Statistics GetStatisticsPolish();
